FINAL UPDATES

+++ Assets/Scripts/SaleItem.cs
@@ -9,7 +9,7 @@ public class SaleItem
     // The game object contains the name
     public GameObject gameObject;
     // The value of the item to buy/sell
-    public int cost;
+    public double cost;
     // The amount of the item for 1 cost
     public int quantity;
     // The amount of the item in stock
@@ -20,7 +20,7 @@ public class SaleItem
     [HideInInspector] public int index;
 
     // Constructor
-    public SaleItem(GameObject gameObject, int cost, int quantity){
+    public SaleItem(GameObject gameObject, double cost, int quantity){
         this.gameObject = gameObject;
         this.cost = cost;
         this.quantity = quantity;
